some initialization fixes
this has a side effect of not showing the description for the synth on the `TrendingSynths` section. if there is a suggeston on how to more easily get this data in a way that is not expensive, please advise. I spent way too long and could not find a better solution for the time being.

* supply the provider configuration to the QueryContext or else it creates its own based on the default
* use `Synths` list from `@synthetixio/contracts-interface` because its statically generated and therefore not expensive.